refractor: move listeners to their own package
